Consolidated DownrangeDistance and ImpactCB factory and parameter classes
